DataWarehouse/pyproject.toml
[project]
name = "datasci-dp"
version = "0.1.0"
description = ""
authors = [
    {name = "narin",email = "narin.w@lexnetix.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "rich (>=14.0.0,<15.0.0)",
    "pandas (>=2.2.3,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
